prod: harden facial-plastics type safety & container build
Mypy strict pass — 103 errors → 0 across 83 source files.

Type fixes (36 files, 294 insertions):
- core/types.py: add missing fields & compat properties
  (Landmark.name, SurfaceMesh.faces alias, RegistrationResult
  rigid_transform/rms_error_mm, MeshQualityReport min/max,
  DicomMetadata voxel_spacing/volume_shape/origin/orientation/hu)
- twin/meshing.py: fix SurfaceMesh constructors (faces→triangles),
  rewrite compute_quality with full field coverage, cap Delaunay
  grid divisions (_MAX_DIVISIONS=20) to prevent O(n²) hang
- twin/registration.py: RegistrationResult field alignment,
  add Modality import
- twin/landmarks.py: Landmark.landmark_type, mesh.triangles
- twin/twin_builder.py: Optional narrowing, variable rename
- data/paired_dataset.py: PopulationSampler API, AnatomyGenerator
  constructor, save_json output
- 22 additional files: no-any-return casts, attr-defined fixes,
  call-arg corrections, assignment alignment

Container build:
- Add products/__init__.py namespace package
- Validate docker build + runtime (34 ops, 15 templates, 12 procs)

Tests: 570/570 pass (36.78s)

+++ products/facial_plastics/sim/aging.py
@@ -119,7 +119,7 @@ class AgingRiskProfile:
         product = 1.0
         for f in factors:
             product *= max(f, 0.01)
-        return product ** (1.0 / len(factors))
+        return float(product ** (1.0 / len(factors)))
 
     def to_dict(self) -> Dict[str, Any]:
         return {
@@ -288,7 +288,7 @@ def _skin_stiffness_evolution(
     collagen_effect = collagen_frac ** 1.5  # nonlinear weakening
     elastin_effect = 0.8 + 0.2 * elastin_frac  # minor contribution
 
-    return collagen_effect * elastin_effect
+    return float(collagen_effect * elastin_effect)
 
 
 def _skin_thickness_evolution(
